@@ -4,7 +4,7 @@
   "description": "Using webpack for javascript sketches",
   "main": "index.js",
   "scripts": {
-    "start:dev": "webpack-dev-server",
+    "start": "webpack-dev-server --hot --history-api-fallback --config webpack.dev.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "release": "echo \"\n\n================\n================\nSTARTING RELEASE BUILD INTO gh-pages BRANCH...\n\" && git checkout gh-pages && git merge master && NODE_ENV=production webpack --optimize-minimize --devtool none && cp public/dist/bundle.app.js ./ && cp public/index.html ./ && ON_DONE_MSG=\"\n\n================\nRELEASED! Ready to commit and push up.\" echo $ON_DONE_MSG"
   },
@@ -20,32 +20,45 @@
   "homepage": "https://github.com/dlopuch/sketchpack",
   "devDependencies": {
     "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.6.0",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
-    "css-loader": "^0.23.1",
-    "eslint": "^4.8.0",
+    "clean-webpack-plugin": "^0.1.17",
+    "copy-webpack-plugin": "^4.5.1",
+    "css-loader": "^0.28.11",
+    "eslint": "^4.12.0",
     "eslint-config-airbnb": "^15.1.0",
-    "eslint-loader": "^1.9.0",
+    "eslint-import-resolver-webpack": "^0.8.3",
+    "eslint-loader": "^2.0.0",
     "eslint-plugin-import": "^2.7.0",
     "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.4.0",
+    "eslint-plugin-react": "^7.7.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^1.1.11",
+    "html-webpack-plugin": "^3.1.0",
     "node-sass": "^4.7.2",
     "postcss-loader": "^2.0.9",
     "sass-loader": "^6.0.6",
     "style-loader": "^0.18.2",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.9.1"
+    "uglifyjs-webpack-plugin": "^1.2.4",
+    "url-loader": "^0.6.2",
+    "webpack": "^4.2.0",
+    "webpack-cli": "^2.0.13",
+    "webpack-dev-server": "^3.1.1",
+    "webpack-merge": "^4.1.2"
   },
   "dependencies": {
     "autoprefixer": "^7.2.1",
-    "bootstrap": "^4.0.0-beta.2",
+    "babel-polyfill": "^6.26.0",
+    "bootstrap": "^4.1.0",
+    "jquery": "^1.9.1",
     "precss": "^2.0.0",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "prop-types": "^15.6.0",
+    "react": "^16.3.1",
+    "react-dom": "^16.3.1",
+    "react-hot-loader": "^4.0.1",
     "react-redux": "^5.0.6",
-    "reactstrap": "^5.0.0-alpha.4",
+    "reactstrap": "^5.0.0",
     "redux": "^3.7.2",
     "redux-actions": "^2.2.1",
     "redux-thunk": "^2.2.0"
